feat(globals): moved globals to globals folder
