+++ src/globals/globals.ts
