+++ src/globals/interfaces.ts
